cloud-function/requirements.txt
functions-framework==3.*
vertexai
google-cloud-aiplatform
pandas
google-cloud-bigquery
pyarrow
